workflow/service_chaining/batch_prediction/requirements.txt
joblib
pandas
scikit-learn
flask>=1.0.2
google-cloud-storage
google-cloud-bigquery
